[vm/aot/tfa] Fix handling of type parameter nullability in factory constructors
TFA represents type parameters inside factory constructors as
additional parameters. In a summary, type parameter type is represented
simply as a reference to a parameter. This approach ignores
nullability of a type parameter type, which is not correct.

This change add a new ApplyNullability operation to a summary in order
to apply any extra nullability ('?' or '*') on top of the type argument
passed to a factory constructor.

TEST=pkg/vm/testcases/transformations/type_flow/transformer/regress_50392_nnbd_strong.dart
Fixes https://github.com/dart-lang/sdk/issues/50392

Change-Id: If2a0d7ea7e100faca23623d4c22b36f3e251b573
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/268381
Reviewed-by: Slava Egorov <vegorov@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>