Flow analysis: simplify `_BranchContext` class.
It turns out the only information stored in `_BranchContext` that was
actually used was the part of `_conditionInfo` that tells what the
flow model should be if the branch is taken.  So we just store that
information directly.

This should make it easier to implement flow analysis for the new
"if-case" construct, which doesn't have a particular expression that's
determining whether the branch is taken (and hence won't have an
easily derivable `_conditionInfo`).

Change-Id: I73ffad36cda4f863335f571542244d58f4bddceb
Bug: https://github.com/dart-lang/sdk/issues/50419
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/268741
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>