[dart2js] Stable sort for entities with same location and name
Sometimes entities have the same name and location. For example, the
synthesized getter and setter for a late instance field have the name
of the field and the location of the original field declaration.  Some
transforms inject multiple closures with the same location and name
('xxx_closure').

We used to 'break ties' by sorting on `hashCode`.  The `hashCode` was
somewhat deterministic as it relied on the `index` of Indexed
entities. The same program unchanged, or under minor changes, tended
to process the entities in the same order and, as the index-derived
hashCode was monotonic in the index, the relative visit order was
invariant when the program was recompiled without changes, and mostly
stable for small changes.

When the deterministic `hashCode` was removed and replaced by the
default `Object.hashCode`, this lead to nondeterministic output, since
the hashCode is different each run.

This could mainly be seen in the sourcemaps where the name might
change between 'foo' and 'foo=' as the relative order of the late
instance field getter and setter changed.


This change removes the comparison on `hashCode`.  Instead, we use a
stable sort, and rely on the entities being sorted having been
collected in some stable order.

Change-Id: I7c8749fea6d2f43bde84bc3991501572f35b1ffa
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/268660
Reviewed-by: Mayank Patke <fishythefish@google.com>
Commit-Queue: Stephen Adams <sra@google.com>