Analyzer: stop using `Index.check` for now.
The analyzer is published on pub, so it needs to confine itself to
APIs that are available in the minimum SDK version specified in its
pubspec.yaml file.  We haven't shipped a version of the SDK that
supports `Index.check` yet, so it's not safe for the analyzer to use
it yet.

Change-Id: I8030869a7548841419fea8d2c3b8f1503a4487eb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/269141
Reviewed-by: Kallen Tu <kallentu@google.com>