[cfe] Use fullMemberIterator in ClassMembersNodeBuilder
This adds support for looking up injected members though the members
builder. It also changes the duplicate handling to match the invariant
used elsewhere; that the used member is the one declared first and not
any of the subsequent, and the so-called, duplicate members.

Change-Id: If9c06a8e17af482b4365da44798b47b54e5277dd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/268946
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>