[vm] Guard against bad brand strings in CPUID
Brand string is supposed to be NULL-terminated.

However we are seeing cases in the wild where
this string is not null-terminated which
causes buffer overrun when trying to print it.

Add trailing '\0' to the brand string to guard
against such cases.

Fixes https://github.com/flutter/flutter/issues/114346

TEST=manually

Change-Id: I21e545bd2fb52336a1fcd7edacb0b867740e8d61
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/269720
Reviewed-by: Martin Kustermann <kustermann@google.com>
Commit-Queue: Slava Egorov <vegorov@google.com>