[cfe] Cache access to scrutinee members in pattern desugaring
The desugaring is presented as a series of nested if-statements
interleaved with intermediate variable declarations that cache the
results of accessing the members of the scrutinee object. Among other
things, it prevents double access to a list element when checking the
pattern conditions on it and then assigning it into a variable pattern.

Part of https://github.com/dart-lang/sdk/issues/49749

Change-Id: I5bf0b040c44d150eacef9a727a8a66fd3a9a568d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/270701
Commit-Queue: Chloe Stefantsova <cstefantsova@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>