[dart2wasm] Fix tear-off getter generation
Currently we use two `hasTearOffUses` and `tornOff` when deciding
whether to generate a tear-off getter for a member.

These metadata are undocumented so it's unclear how they are supposed to
be used, but `hasTearOfUses` seems to be what we need when generating
references for tear-off getters. For example, in this code:

    class A {
      void f() => print('A.f called');
    }

    void main() {
      dynamic x = A();
      final f = x.f;
      f();
    }

`hasTearOffUses` of `A.f` is true, `tornOff` is false. So the current
condition `hasTearOffUses && tornOff` incorrectly skips the reference.

In addition, I realized in my testing that when `tornOff` is true,
`hasTearOffUses` is true as well. So just using `hasTearOffUses` seems
to do the right thing.

New passing tests:

- corelib/map_unmodifiable_cast_test
- language/closure/bound_closure_equality_test
- language/closure/closure7_test
- language/closure/tearoff_bounds_instantiation_test
- language/function/type3_test
- language/nosuchmethod_forwarding/nosuchmethod_forwarding_test/06
- language/prefix/prefix15_test
- language/unsorted/fast_method_extraction_test
- lib/async/stream_event_transformed_test
- lib/convert/chunked_conversion1_test
- lib/convert/streamed_conversion_json_decode1_test
- lib/convert/streamed_conversion_json_utf8_decode_test

Fixes #50485

Change-Id: Id130649d6e175b84430f9bf967ea01c4d471cd92
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/271320
Reviewed-by: Aske Simon Christensen <askesc@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>