Leave `dependencies:` node uncommented
Ideally we'd leave the `dependencies:` node itself uncommented, even if the content beneath it is commented out. If the node is commented, then VSCode's "Dart: Add Dependency" command gets confused, and adds a new node at the bottom, rather than uncommenting the `dependencies:` node. 

This change should be benign, and removes the need for Dart-Code to introduce logic to uncomment the parent node. For finesse, this also adds an explanatory comment to indicate that `path:` is an example.